refactor: shift components from tracklist into they're own files
