+++ src/screens/Details/TrackList/Tracklist.styles.ts
@@ -1,20 +1,15 @@
-import { ImageStyle, StyleSheet, TextStyle, ViewStyle } from 'react-native';
+import { StyleSheet, TextStyle, ViewStyle } from 'react-native';
 import theme from '../../../theme';
 
 type Style = {
   artistName: TextStyle;
   noSamples: TextStyle;
   noSamplesWrapper: ViewStyle;
-  skeletonLoader: ViewStyle;
   snackBar: ViewStyle;
-  trackImage: ImageStyle;
-  trackItem: ViewStyle;
-  trackListWrapper: ViewStyle;
-  trackName: TextStyle;
   view: ViewStyle;
 };
 
-const { defaultPalette, padding, borderRadius, fontSizes } = theme;
+const { defaultPalette, padding, fontSizes } = theme;
 
 export default StyleSheet.create<Style>({
   artistName: {
@@ -28,31 +23,8 @@ export default StyleSheet.create<Style>({
     alignItems: 'center',
     justifyContent: 'center',
   },
-  skeletonLoader: {
-    flex: 1,
-    height: 275,
-  },
   snackBar: {
     backgroundColor: defaultPalette.success,
-  },
-  trackImage: {
-    borderRadius: borderRadius / 4,
-  },
-  trackItem: {
-    marginVertical: '3%',
-    width: '80%',
-    flex: 1,
-    flexDirection: 'row',
-    padding: padding.base,
-    backgroundColor: defaultPalette.secondaryBackground,
-    borderRadius,
-  },
-  trackListWrapper: {
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  trackName: {
-    fontSize: fontSizes.MEDIUM,
   },
   view: {
     width: '100%',
